+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-13T03:10:24.062Z",
+    "2026-07-13T04:38:02.412Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {