+++ packages/orq-rc/RELEASES.md
@@ -10719,3 +10719,13 @@ Based on:
 - [typescript v4.12.0-rc.16] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.16 - packages/orq-rc
+
+## 2026-07-13 04:40:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.3 (2.916.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.17] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.17 - packages/orq-rc